requirements.txt
selenium
python-dotenv
requests
bs4
openai
lxml